+++ recon/named/ext_flash_api_init.c
@@ -23,10 +23,10 @@
 #include "../headers/g1_log.h"
 /* Reconstructed ext_flash_api_init @ 0x2e4e8  (parity: 300/300 trials, PROVEN) */
 
-extern void kmutex_dlist_init(unsigned int);
-extern unsigned int z_device_is_ready(unsigned int);
+extern void kmutex_dlist_init(unsigned long);
+extern int z_device_is_ready(int);
 extern unsigned short z_impl_flash_get_page_count(unsigned int);
-extern unsigned int get_device_info(void);
+extern int get_device_info(void);
 
 unsigned int ext_flash_api_init(unsigned int *param_1, unsigned int *param_2, unsigned int *param_3, unsigned int param_4)
 {
