+++ recon/named/gpio_configure_pin_output.c
@@ -13,7 +13,7 @@
 #include "../headers/g1_nrf_regs.h"
 #include "../headers/g1_log.h"
 /* Reconstructed FUN_0006666c @ 0x6666c  (parity: 300/300 trials, PROVEN) */
-extern long long assert_post_action(unsigned,unsigned);
+extern int assert_post_action(int, int);
 void gpio_configure_pin_output(unsigned param_1, int param_2, unsigned param_3, unsigned param_4){
     unsigned r0v = param_1;
     unsigned r1v = (unsigned)param_2;
