+++ recon/named/net_buf_simple_push.c
@@ -13,7 +13,7 @@
 #include "../headers/g1_log.h"
 /* Reconstructed FUN_0005f518 @ 0x5f518  (parity: 300/300 trials, PROVEN) */
 
-extern unsigned long long assert_post_action(unsigned int, unsigned int);
+extern int assert_post_action(int, int);
 
 void net_buf_simple_push(int *param_1, unsigned int param_2, unsigned int param_3, unsigned int param_4)
 {
