+++ recon/named/nrfx_gppi_task_endpoint_clear.c
@@ -14,7 +14,7 @@
  * Raw identity/back-map: FUN_00064fd4.  Parity: 300/300 trials, PROVEN.
  */
 
-extern unsigned long long assert_post_action(unsigned int,unsigned int);
+extern int assert_post_action(int, int);
 void nrfx_gppi_task_endpoint_clear(
     unsigned int channel, int task_endpoint,
     unsigned int param_3, unsigned int param_4)
