+++ recon/readable_sources/app/g1/display_panel_is_secondary.c
@@ -13,7 +13,7 @@
 /* Reconstructed check_battery_critical @ 0x2bed0  (parity: 300/300 trials, PROVEN) */
 
 extern void signal_persist_task_event(void);
-unsigned int display_panel_is_secondary(int param_1){
+int display_panel_is_secondary(unsigned long param_1) {
     volatile unsigned char *pcVar1 = (volatile unsigned char*)((unsigned long)&g_persist_task_status_lock) /*=0x20018d9c*/;
     while(*pcVar1 != 0){
         signal_persist_task_event();
